+++ src/QueryToMany.php
@@ -28,9 +28,10 @@ use Illuminate\Database\Eloquent\Relations\Relation as EloquentRelation;
 use Illuminate\Support\Collection;
 use Illuminate\Support\LazyCollection;
 use LaravelJsonApi\Contracts\Pagination\Page;
-use LaravelJsonApi\Contracts\Query\QueryParameters;
+use LaravelJsonApi\Contracts\Query\QueryParameters as QueryParametersContract;
 use LaravelJsonApi\Contracts\Store\QueryManyBuilder;
 use LaravelJsonApi\Core\Query\IncludePaths;
+use LaravelJsonApi\Core\Query\QueryParameters;
 use LaravelJsonApi\Core\Query\SortFields;
 use LaravelJsonApi\Eloquent\Fields\Relations\ToMany;
 use LogicException;
@@ -39,6 +40,8 @@ use function sprintf;
 
 class QueryToMany implements QueryManyBuilder
 {
+
+    use HasQueryParameters;
 
     /**
      * @var Model
@@ -51,21 +54,6 @@ class QueryToMany implements QueryManyBuilder
     private ToMany $relation;
 
     /**
-     * @var array|null
-     */
-    private ?array $filters = null;
-
-    /**
-     * @var SortFields|null
-     */
-    private ?SortFields $sort = null;
-
-    /**
-     * @var IncludePaths|null
-     */
-    private ?IncludePaths $includePaths = null;
-
-    /**
      * QueryToMany constructor.
      *
      * @param Model $model
@@ -75,17 +63,7 @@ class QueryToMany implements QueryManyBuilder
     {
         $this->model = $model;
         $this->relation = $relation;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function using(QueryParameters $query): QueryManyBuilder
-    {
-        return $this
-            ->with($query->includePaths())
-            ->filter($query->filter())
-            ->sort($query->sortFields());
+        $this->queryParameters = new QueryParameters();
     }
 
     /**
@@ -93,7 +71,7 @@ class QueryToMany implements QueryManyBuilder
      */
     public function filter(?array $filters): QueryManyBuilder
     {
-        $this->filters = $filters;
+        $this->queryParameters->setFilters($filters);
 
         return $this;
     }
@@ -103,17 +81,7 @@ class QueryToMany implements QueryManyBuilder
      */
     public function sort($fields): QueryManyBuilder
     {
-        $this->sort = SortFields::nullable($fields);
-
-        return $this;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function with($includePaths): QueryManyBuilder
-    {
-        $this->includePaths = IncludePaths::nullable($includePaths);
+        $this->queryParameters->setSortFields($fields);
 
         return $this;
     }
@@ -123,7 +91,7 @@ class QueryToMany implements QueryManyBuilder
      */
     public function get(): Collection
     {
-        return $this->prepareQuery()->get();
+        return $this->query()->get();
     }
 
     /**
@@ -131,7 +99,7 @@ class QueryToMany implements QueryManyBuilder
      */
     public function cursor(): LazyCollection
     {
-        return $this->prepareQuery()->cursor();
+        return $this->query()->cursor();
     }
 
     /**
@@ -139,7 +107,7 @@ class QueryToMany implements QueryManyBuilder
      */
     public function paginate(array $page): Page
     {
-        return $this->prepareQuery()->paginate($page);
+        return $this->query()->paginate($page);
     }
 
     /**
@@ -163,22 +131,15 @@ class QueryToMany implements QueryManyBuilder
      */
     public function query(): Builder
     {
-        return new Builder(
+        $query = new Builder(
             $this->relation->schema(),
             $this->getRelation(),
             $this->relation
         );
-    }
 
-    /**
-     * @return Builder
-     */
-    private function prepareQuery(): Builder
-    {
-        return $this->query()
-            ->with($this->includePaths)
-            ->filter($this->filters)
-            ->sort($this->sort);
+        $query->withQueryParameters($this->queryParameters);
+
+        return $query;
     }
 
     /**